+++ psycop/projects/forced_admission_inpatient/model_eval/run_pipeline_on_val.py
@@ -55,7 +55,7 @@ def _train_pipeline_on_test(
     cfg.project.wandb.Config.allow_mutation = True
     cfg.data.Config.allow_mutation = True
     cfg.data.splits_for_training = splits_for_training
-    cfg.data.datasets_for_evaluation = datasets_for_evaluation
+    cfg.data.splits_for_evaluation = datasets_for_evaluation
 
     override_dir = _get_test_pipeline_dir(pipeline_to_train=pipeline_to_train)
     msg.info(f"Evaluating to {override_dir}")