+++ psycop/projects/t2d/paper_outputs/model_description/feature_importance/gain/gain_table.py
@@ -3,7 +3,7 @@ import polars as pl
 from psycop.common.model_training.data_loader.utils import (
     load_and_filter_split_from_cfg,
 )
-from psycop.projects.t2d.paper_outputs.selected_runs import BEST_EVAL_PIPELINE
+from psycop.projects.t2d.paper_outputs.selected_runs import get_best_eval_pipeline
 from psycop.projects.t2d.utils.feature_name_to_readable import feature_name_to_readable
 from psycop.projects.t2d.utils.pipeline_objects import T2DPipelineRun
 
@@ -58,6 +58,6 @@ def generate_feature_importance_table(pipeline_run: T2DPipelineRun) -> pl.DataFr
 
 if __name__ == "__main__":
     top_100_features = generate_feature_importance_table(
-        pipeline_run=BEST_EVAL_PIPELINE,
+        pipeline_run=get_best_eval_pipeline,
     )
     pass