+++ psycop/projects/t2d/paper_outputs/model_description/performance/confusion_matrix_pipeline.py
@@ -34,6 +34,6 @@ def t2d_confusion_matrix_plot(run: T2DPipelineRun) -> pn.ggplot:
 
 
 if __name__ == "__main__":
-    from psycop.projects.t2d.paper_outputs.selected_runs import BEST_EVAL_PIPELINE
+    from psycop.projects.t2d.paper_outputs.selected_runs import get_best_eval_pipeline
 
-    t2d_confusion_matrix_plot(run=BEST_EVAL_PIPELINE)
+    t2d_confusion_matrix_plot(run=get_best_eval_pipeline)