+++ psycop/projects/t2d/paper_outputs/model_description/performance/main_performance_figure.py
@@ -30,6 +30,6 @@ def t2d_create_main_performance_figure(run: T2DPipelineRun) -> None:
 
 
 if __name__ == "__main__":
-    from psycop.projects.t2d.paper_outputs.selected_runs import BEST_EVAL_PIPELINE
+    from psycop.projects.t2d.paper_outputs.selected_runs import get_best_eval_pipeline
 
-    t2d_create_main_performance_figure(run=BEST_EVAL_PIPELINE)
+    t2d_create_main_performance_figure(run=get_best_eval_pipeline)