+++ psycop/projects/t2d/paper_outputs/model_description/robustness/robustness_by_cyclic_time.py
@@ -6,7 +6,7 @@ from psycop.common.model_evaluation.binary.time.periodic_data import (
 from psycop.projects.t2d.paper_outputs.model_description.robustness.robustness_plot import (
     t2d_plot_robustness,
 )
-from psycop.projects.t2d.paper_outputs.selected_runs import BEST_EVAL_PIPELINE
+from psycop.projects.t2d.paper_outputs.selected_runs import get_best_eval_pipeline
 from psycop.projects.t2d.utils.pipeline_objects import T2DPipelineRun
 
 
@@ -47,5 +47,5 @@ def t2d_auroc_by_month_of_year(run: T2DPipelineRun) -> pn.ggplot:
 
 
 if __name__ == "__main__":
-    t2d_auroc_by_day_of_week(run=BEST_EVAL_PIPELINE)
-    t2d_auroc_by_month_of_year(run=BEST_EVAL_PIPELINE)
+    t2d_auroc_by_day_of_week(run=get_best_eval_pipeline)
+    t2d_auroc_by_month_of_year(run=get_best_eval_pipeline)