+++ psycop/projects/t2d/paper_outputs/model_permutation/snoozing/try_snoozing.py
@@ -6,10 +6,10 @@ from sklearn.metrics import roc_auc_score
 from psycop.projects.t2d.paper_outputs.model_permutation.snoozing.snoozing import (
     snooze_dataframe,
 )
-from psycop.projects.t2d.paper_outputs.selected_runs import BEST_EVAL_PIPELINE
+from psycop.projects.t2d.paper_outputs.selected_runs import get_best_eval_pipeline
 
 if __name__ == "__main__":
-    evaluation_dataset = BEST_EVAL_PIPELINE.pipeline_outputs.get_eval_dataset()
+    evaluation_dataset = get_best_eval_pipeline.pipeline_outputs.get_eval_dataset()
 
     eval_df = pd.DataFrame(
         {