+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ### Fix
 
 * Change typehint for patient colnames ([`22d9317`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/22d931731fc96d42e8ced0671e63a44be9dcfad5))
-* Do not import BEST_EVAL_PIPELINE unless main ([`d5da51f`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d5da51fbf9c66e1ef5c75d42657f1002b56a3605))
+* Do not import get_best_eval_pipeline unless main ([`d5da51f`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/d5da51fbf9c66e1ef5c75d42657f1002b56a3605))
 * Fixed mutable default error in config ([`a2d8294`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/a2d8294f62c4b19a6b2801ce9a9898585deb4327))
 * Source subtype filtering works ([`1259203`](https://github.com/Aarhus-Psychiatry-Research/psycop-common/commit/1259203d58b09ee2b09f71f7455976199fc9ed10))